Added New css and images and modify  parent slider and add parent component
