+++ ChildHome/public/index.html
@@ -45,7 +45,13 @@
 
     <title>HomeForHeart</title>
   </head>
-  <body>
+  <body
+    style="
+      background-image: url('/images/bg.jpg');
+      background-size: cover;
+      background-repeat: no-repeat;
+    "
+  >
     <div id="root"></div>
   </body>
 </html>
