+++ ChildHome/src/Components/Donation/RazorpayPayment.css
@@ -1,7 +1,5 @@
 .payment-card {
     border-radius: 15px;
-    background: linear-gradient(135deg, #f6d365 0%, #fda085 100%);
-    color: #fff;
   }
   
   .payment-card .card-title {
@@ -16,7 +14,7 @@
   }
   
   .payment-card .pay-button {
-    background-color: #ff6f61;
+    background-color: #619eff;
     border: none;
     border-radius: 10px;
     padding: 10px;
@@ -29,5 +27,5 @@
   }
   
   .payment-card .text-muted {
-    color: #e0e0e0 !important;
+    color: #000000 !important;
   }
