+++ ChildHome/src/Components/Registration/Registration.css
@@ -2,7 +2,7 @@
 
 .container_background{
     /* background: linear-gradient(135deg, #ffffff, #ff6a88, #ff99ac); */
-    background-color: white;
+   
     padding: 20px;
   }
 .registration-container {
